terraform_azure_backend/requirements.txt
Django==5.1.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
